requirements.txt
pymupdf>=1.24.0
openai>=1.0.0
litellm>=1.0.0
pyyaml>=6.0.0
